Copy platform from baremetal app
