+++ src/platform/imx8/plat.mk
@@ -1,2 +1,3 @@
 ARCH:=armv8
 GIC_VERSION:=GICV3
+drivers:=nxp_uart
