+++ src/platform/imx8/sources.mk
@@ -1,2 +1,2 @@
-plat_c_srcs:= nxp_uart.c
+plat_c_srcs:=imx8.c
 plat_s_srcs:=
