+++ src/platform/qemu/aarch64-virt/plat.mk
@@ -1,2 +1,3 @@
 ARCH:=armv8
 GIC_VERSION:=GICV3
+drivers:=pl011_uart
