+++ src/platform/qemu/riscv64-virt/plat.mk
@@ -1,1 +1,2 @@
 ARCH:=riscv
+drivers:=8250_uart
