+++ src/platform/qemu/riscv64-virt/sources.mk
@@ -1,2 +1,2 @@
-plat_c_srcs:= uart8250.c
+plat_c_srcs:=virt.c
 plat_s_srcs:=
