+++ src/platform/tx2/inc/plat.h
@@ -1,0 +1,9 @@
+#ifndef PLAT_H
+#define PLAT_H
+
+#define PLAT_MEM_BASE 0x0
+#define PLAT_MEM_SIZE 0x1000000
+
+#define UART_IRQ_ID (144)
+
+#endif
