+++ src/platform/tx2/plat.mk
@@ -1,3 +1,3 @@
 ARCH:=armv8
 GIC_VERSION:=GICV2
-drivers:=zynq_uart
+drivers:=8250_uart
