+++ src/platform/tx2/sources.mk
@@ -1,0 +1,2 @@
+plat_c_srcs:=tx2.c
+plat_s_srcs:=
